+++ User Interface/VoiceOptimizer/VoiceOptimizer/Assets.xcassets/Speak.imageset/Contents.json
@@ -2,8 +2,18 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "Voice Recognition Scan Filled-50.png",
+      "filename" : "Microphone-96.png",
       "scale" : "1x"
+    },
+    {
+      "idiom" : "universal",
+      "filename" : "Microphone-97.png",
+      "scale" : "2x"
+    },
+    {
+      "idiom" : "universal",
+      "filename" : "Microphone-98.png",
+      "scale" : "3x"
     }
   ],
   "info" : {
