+++ User Interface/VoiceOptimizer/VoiceOptimizer/Assets.xcassets/Stop.imageset/Contents.json
@@ -2,15 +2,17 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "Play-50.png",
+      "filename" : "Microphone-96 (1).png",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
+      "filename" : "Microphone-96 (1)-1.png",
       "scale" : "2x"
     },
     {
       "idiom" : "universal",
+      "filename" : "Microphone-96 (1)-2.png",
       "scale" : "3x"
     }
   ],
